module5-task2 Доработка формы
