+++ js/filters.js
@@ -1,18 +1,8 @@
 import { typeSelect, priceSelect, roomsSelect, guestsSelect, featuresFieldset } from './utils.js';
+import {getPrice} from './helpers.js';
 
 
 const Filters = (data) => {
-
-  const getPrice = (price) => {
-    switch (price) {
-      case 'low':
-        return {min: 0, max: 10000};
-      case 'middle':
-        return {min: 10000, max: 50000};
-      case 'high':
-        return {min: 50000, max: Infinity};
-    }
-  }
 
   const filtrationItem = (el, item, param) => {
     if(el.value === 'any') {
